scoring/requirements.txt
confluent-kafka==1.9.2
protobuf==3.20.*
levenshtein==0.20.8
tldextract==3.4.0
PyYAML==6.0
requests==2.28.1